Dog vs Cat Neural network model
This code showcases an exciting machine-learning project using Keras to build a highly accurate image classification model for differentiating between cats and dogs. We explore visual effects, preprocess the dataset, design a thoughtful neural network, and ensure its performance through monitoring and early stopping. The model's success is evident in its ability to predict new images accurately. With responsible preparation, I save the model for future use in various applications. This project basically tells the model to understand and tell the difference in pictures of which picture is a cat or a dog classifier. 

here is the dataset:
https://www.kaggle.com/competitions/dogs-vs-cats/data